feat: 에피소드 조회 기능 추가

+++ src/main/java/pretzel/dreamketcherbe/domain/episode/entity/Episode.java
@@ -48,7 +48,7 @@ public class Episode extends BaseTimeEntity {
 
     @ColumnDefault("0")
     @Column(name = "view_count")
-    private Long viewCount;
+    private int viewCount;
 
     @ManyToOne
     @JoinColumn(name = "webtoon_id")
